# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tripio",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/requests": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "Вывод списка записей (index())",
                "description": "Доступ только авторизованным пользователям",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    },
                    "200": {
                        "description": "Everything OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RequestsSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Добавление новой записи (store())",
                "description": "Возвращает новый созданный route",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RequestsSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/requests/{id}": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "Вывод конкретной записи (show())",
                "description": "Returns request data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Route id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RequestsSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Requests"
                ],
                "summary": "Обновление записи (update())",
                "description": "Returns updated request data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestsSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RequestsSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Requests"
                ],
                "summary": "Удаление записи (delete())",
                "description": "Deletes a record and returns no content",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RequestsSchema": {
                "properties": {
                    "id": {
                        "description": "Request ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "service_id": {
                        "description": "Foreign key service",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "client_id": {
                        "description": "Foreign key service",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "client_phone": {
                        "description": "Номер пользователя, кто оставил заявку на экскурсию",
                        "type": "string",
                        "example": "9786666666",
                        "nullable": false
                    },
                    "client_name": {
                        "description": "Имя пользователя, кто оставил заявку на экскурсию",
                        "type": "string",
                        "example": "Арина",
                        "nullable": false
                    },
                    "message": {
                        "description": "Сообщение пользователя",
                        "type": "text",
                        "example": "Хотим попасть к вам на экскурсию по Казани",
                        "nullable": false
                    },
                    "created_at": {
                        "description": "Date of route creation",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    },
                    "updated_at": {
                        "description": "Date of last updating route data",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}